# web/src/api/openapi.merged.yaml
openapi: 3.1.0
info:
  title: ATRAG API
  description: ATRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
  /bots:
    get:
      summary: List bots
      description: Get a list of bots
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of bots
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/botList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    post:
      summary: Create a new bot
      description: Create a new bot
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/botCreate'
      responses:
        '200':
          description: Bot created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bot'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /bots/{bot_id}:
    get:
      summary: Get bot details
      description: Get details of a specific bot
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bot details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/bot'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    delete:
      summary: Delete a bot
      description: Delete a bot (idempotent operation)
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Bot deleted successfully (or already deleted)
    put:
      summary: Update a bot
      description: Update a bot
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/botUpdate'
      responses:
        '200':
          description: Bot updated successfully
...
  /bots/{bot_id}/flow/debug:
    post:
      summary: Debug a flow
      description: Debug a flow
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/debugFlowRequest'
      responses:
        '200':
          description: Flow debugged successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /bots/{bot_id}/chats:
    get:
      summary: List chats
      description: Get a paginated list of chats
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number (1-based)
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Number of items per page
      responses:
        '200':
          description: Paginated list of chats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chatList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    post:
      summary: Create a new chat
      description: Create a new chat
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chatCreate'
      responses:
        '201':
          description: Chat created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chat'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /bots/{bot_id}/chats/{chat_id}:
    get:
      summary: Get chat details
      description: Get details of a specific chat
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chat details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chatDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    put:
      summary: Update a chat
      description: Update a chat
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/chatUpdate'
      responses:
        '200':
          description: Chat updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/chat'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
    delete:
      summary: Delete a chat
      description: Delete a chat (idempotent operation)
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Chat deleted successfully (or already deleted)
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /bots/{bot_id}/chats/{chat_id}/title:
    post:
      summary: Generate a chat title
      description: Generate a short title for the chat based on recent conversation content
      security:
        - BearerAuth: []
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/titleGenerateRequest'
      responses:
        '200':
          description: Generated title
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/titleGenerateResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '404':
          description: Chat not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
  /bots/{bot_id}/chats/{chat_id}/messages/{message_id}:
    post:
      summary: Feedback a message
      description: Feedback a message
      parameters:
        - name: bot_id
          in: path
          required: true
          schema:
            type: string
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
        - name: message_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/feedback'
      responses:
        '200':
          description: A message
  /chats/{chat_id}/documents:
    post:
      summary: Upload document to chat
      description: Upload a document file to a chat session for temporary knowledge augmentation
      tags:
        - chat-documents
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: Chat session ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file to upload
                message_id:
                  type: string
                  description: Message ID to associate the document with
              required:
                - file
                - message_id
      responses:
        '200':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/document'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Chat not found
  /chats/{chat_id}/documents/{document_id}:
    get:
      summary: Get chat document details
      description: Get details of a specific document in a chat session
      tags:
        - chat-documents
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
          description: Chat session ID
        - name: document_id
          in: path
          required: true
          schema:
            type: string
          description: Document ID
      responses:
        '200':
          description: Document details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/document'
        '401':
          description: Unauthorized
        '404':
          description: Document or chat not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
    modelSpec:
      type: object
      properties:
        model:
          type: string
          description: The name of the language model to use
          example: gpt-4o-mini
        model_service_provider:
          type: string
          description: Used for querying auth information (api_key/api_base/...) for a model service provider.
          example: openai
        custom_llm_provider:
          type: string
          description: Used for Non-OpenAI LLMs (e.g. 'bedrock' for amazon.titan-tg1-large)
          example: openai
        temperature:
          type: number
          description: Controls randomness in the output. Values between 0 and 2. Lower values make output more focused and deterministic
          minimum: 0
          maximum: 2
          example: 0.1
          default: 0.1
        max_tokens:
          type: integer
          description: Maximum number of tokens to generate
          minimum: 1
          example: 4096
        max_completion_tokens:
          type: integer
          description: Upper bound for generated completion tokens, including visible and reasoning tokens
          minimum: 1
          example: 4096
        timeout:
          type: integer
          description: Maximum execution time in seconds for the API request
          minimum: 1
        top_n:
          type: integer
          description: Number of top results to return when reranking documents
          minimum: 1
        tags:
          type: array
          items:
            type: string
          description: Tags for model categorization
          example:
            - free
            - recommend
          default: []
...
    botConfig:
      type: object
      properties:
        agent:
          type: object
          properties:
            completion:
              $ref: '#/components/schemas/modelSpec'
            routing_completion:
              $ref: '#/components/schemas/modelSpec'
              description: Optional default routing LLM. Its API key is resolved from the bot owner's provider configuration.
            system_prompt_template:
              type: string
            query_prompt_template:
              type: string
            collections:
              type: array
              items:
                $ref: '#/components/schemas/collection'
        flow:
          $ref: '#/components/schemas/WorkflowDefinition'
    bot:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
            - knowledge
            - common
            - agent
          description: The type of bot
          example: knowledge
        config:
          $ref: '#/components/schemas/botConfig'
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
    pageResult:
      description: PageResult info (deprecated, use paginatedResponse instead)
      properties:
        page_number:
          type: integer
          description: The page number
        page_size:
          type: integer
          description: The page size
        count:
          type: integer
          description: The total count of items
      type: object
    botList:
      type: object
      description: A list of bots
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/bot'
        pageResult:
          $ref: '#/components/schemas/pageResult'
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
    botCreate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum:
            - knowledge
            - common
            - agent
          description: The type of bot
          example: knowledge
        config:
          $ref: '#/components/schemas/botConfig'
    botUpdate:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        config:
          $ref: '#/components/schemas/botConfig'
    debugFlowRequest:
      type: object
      properties:
        query:
          type: string
      required:
        - query
    chat:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        bot_id:
          type: string
        peer_id:
          type: string
        peer_type:
          type: string
          enum:
            - system
            - feishu
            - weixin
            - weixin_official
            - web
            - dingtalk
        status:
          type: string
          enum:
            - active
            - archived
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
    paginatedResponse:
      type: object
      properties:
        total:
          type: integer
          minimum: 0
          description: Total number of items
          example: 100
        page:
          type: integer
          minimum: 1
          description: Current page number
          example: 1
        page_size:
          type: integer
          minimum: 1
          description: Number of items per page
          example: 10
        total_pages:
          type: integer
          minimum: 1
          description: Total number of pages
          example: 10
        has_next:
          type: boolean
          description: Whether there is a next page
          example: true
        has_prev:
          type: boolean
          description: Whether there is a previous page
          example: false
    chatList:
      type: object
      description: A list of chats with pagination
      allOf:
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/chat'
        - $ref: '#/components/schemas/paginatedResponse'
    chatCreate:
      type: object
      properties:
        title:
          type: string
    reference:
      type: object
      properties:
        score:
          type: number
        text:
          type: string
        image_uri:
          type: string
        metadata:
          type: object
          additionalProperties: true
    feedback:
      type: object
      properties:
        type:
          type: string
          enum:
            - good
            - bad
        tag:
          type: string
          enum:
            - Harmful
            - Unsafe
            - Fake
            - Unhelpful
            - Other
        message:
          type: string
    file:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
    chatMessage:
      type: object
      properties:
        id:
          type: string
        part_id:
          type: string
        type:
          type: string
          enum:
            - welcome
            - message
            - start
            - stop
            - error
            - tool_call_result
            - thinking
            - references
        timestamp:
          type: number
        role:
          type: string
          enum:
            - human
            - ai
        data:
          type: string
        references:
          type: array
          items:
            $ref: '#/components/schemas/reference'
        urls:
          type: array
          items:
            type: string
        feedback:
          $ref: '#/components/schemas/feedback'
        files:
          type: array
          items:
            $ref: '#/components/schemas/file'
    chatDetails:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        bot_id:
          type: string
        peer_id:
          type: string
        peer_type:
          type: string
          enum:
            - system
            - feishu
            - weixin
            - weixin_official
            - web
            - dingtalk
        history:
          type: array
          description: Array of conversation turns, where each turn is an array of message parts
          items:
            type: array
            description: A conversation turn containing multiple message parts (thinking, content, references, etc.)
            items:
              $ref: '#/components/schemas/chatMessage'
        status:
          type: string
          enum:
            - active
            - archived
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
    chatUpdate:
      type: object
      properties:
        title:
          type: string
    titleGenerateRequest:
      type: object
      properties:
        max_length:
          type: integer
          description: Maximum length of the generated title
          default: 20
          minimum: 6
          maximum: 50
        language:
          type: string
          description: Language for the title generation (IETF BCP 47 tag)
          default: zh-CN
          enum:
            - zh-CN
            - en-US
            - ja-JP
            - ko-KR
        turns:
          type: integer
          description: Number of most recent conversation turns to consider
          default: 1
          minimum: 1
    titleGenerateResponse:
      type: object
      properties:
        title:
          type: string
          description: Generated title string
      required:
        - title
    document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
            - UPLOADED
            - EXPIRED
            - PENDING
            - RUNNING
            - COMPLETE
            - FAILED
            - DELETING
            - DELETED
        vector_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        fulltext_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        graph_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        summary_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        vision_index_status:
          type: string
          enum:
            - PENDING
            - CREATING
            - ACTIVE
            - DELETING
            - DELETION_IN_PROGRESS
            - FAILED
            - SKIPPED
        vector_index_updated:
          type: string
          format: date-time
          description: Vector index last updated time
        fulltext_index_updated:
          type: string
          format: date-time
          description: Fulltext index last updated time
        graph_index_updated:
          type: string
          format: date-time
          description: Graph index last updated time
        summary_index_updated:
          type: string
          format: date-time
          description: Summary index last updated time
        vision_index_updated:
          type: string
          format: date-time
          description: Vision index last updated time
        summary:
          type: string
          description: Summary of the document
        size:
          type: number
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
...